@@ -1,6 +1,6 @@
 {
   "name": "@phalgunv/ngx-keyboard-shortcuts",
-  "version": "17.0.0",
+  "version": "18.0.0",
   "private": true,
   "scripts": {
     "start": "npm run build && npm run copy-package-to-demo-node-modules && ng serve",
@@ -41,23 +41,23 @@
     "url": "https://github.com/phalgunv/ngx-keyboard-shortcuts/issues"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.17",
-    "@angular-eslint/builder": "^17.5.2",
-    "@angular-eslint/eslint-plugin": "^17.5.2",
-    "@angular-eslint/eslint-plugin-template": "^17.5.2",
-    "@angular-eslint/schematics": "^17.5.2",
-    "@angular-eslint/template-parser": "^17.5.2",
-    "@angular/animations": "^17.3.12",
-    "@angular/cli": "^17.3.17",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/language-service": "^17.3.12",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.21",
+    "@angular-eslint/builder": "^18.4.2",
+    "@angular-eslint/eslint-plugin": "^18.4.2",
+    "@angular-eslint/eslint-plugin-template": "^18.4.2",
+    "@angular-eslint/schematics": "^18.4.2",
+    "@angular-eslint/template-parser": "^18.4.2",
+    "@angular/animations": "^18.2.14",
+    "@angular/cli": "^18.2.21",
+    "@angular/common": "^18.2.14",
+    "@angular/compiler": "^18.2.14",
+    "@angular/compiler-cli": "^18.2.14",
+    "@angular/core": "^18.2.14",
+    "@angular/forms": "^18.2.14",
+    "@angular/language-service": "^18.2.14",
+    "@angular/platform-browser": "^18.2.14",
+    "@angular/platform-browser-dynamic": "^18.2.14",
+    "@angular/router": "^18.2.14",
     "@types/jest": "^29.5.3",
     "@typescript-eslint/eslint-plugin": "^8.46.2",
     "@typescript-eslint/parser": "^8.46.2",
@@ -69,13 +69,13 @@
     "jest": "^29.7.0",
     "jest-junit": "^12.0.1",
     "jest-preset-angular": "^13.0.0",
-    "ng-packagr": "^17.3.0",
+    "ng-packagr": "^18.2.1",
     "ngx-highlightjs": "^4.1.2",
     "puppeteer": "^24.26.1",
     "reflect-metadata": "^0.1.13",
     "rxjs": "~7.8.0",
     "ts-jest": "^29.1.0",
-    "typescript": "5.4.5",
+    "typescript": "5.5.4",
     "zone.js": "^0.14.10"
   },
   "dependencies": {
